Changes to support a few use cases
1. Allow model, tmpl create in a view to happen, give _init a chance to
make changes to them before they are applied to the view.
2. Added extra logic in resources to allow defining resources in the
form of an array.  Great for shorthand notation of dependencies
computed from the FQN passed in.
3. Added the ability define a model in a view as a direct property or a
function, which gets resolved when loading resources.
4. Added logic in resources to figure where to find resources based on
the type of resource being loaded.  E.g. model will be found in the
model folder, tmpl will be found in the tmpl folder and so on.
5. Added logic in resources to allow the addition of custom resource
handlers